Fix for chapter number duplication by author.
The author misnumbered chapter 1633 as 1632, then continued numbering
from there. When it was noticed, many chapters would have needed
renumbering, so he did a "quick & dirty" fix.

If this isn't taken into consideration, one of the two chapters labeled
as 1632 will be missing from the epub, so here is a workaround.