Change how WuxiaWorld chapter titles are processed
The old method would take the parts between the first and third " - "
cuts, the new one takes the parts between the first and last cut.

In cases where the chapter title includes a " - " to add a "second
title" like "HJC - Chapter 288 - The Saint Land - The Peerless Sect (3)
- WuxiaWorld", it would end as "Chapter 288 - The Saint Land", missing
entirely the " - The Peerless Sect (3)", resulting in 3 chapters with
the same exact name (the 3 is an index for a multipart chapter).

This may also affect other titles with a lesser impact (just a longer
title).