Attempt to create our own custom user agent.
The previous user agent is getting old and with Google phasing out
support android for 4.x things will get worse sooner or later, so we
need to switch to one which will work for us in the long term.

Some servers may try to block our scrapping of their novels by rejecting
this custom user agent, and that would force us to fake it at some
point, but at least for now it should work.