Small fix for cobi's chapter list generation regex
The previous regexp worked well before the chapter list grew, and to get
newer volumes, but failed to regenerate the chapter list for older
volumes.

Looking for chapter names starting with "1" followed by any chapter
title made it find chapters "1", "10", "11", ..., "100", "101", ...,
"1000", ..., and then settled on the last one found (currently 1359, but
eventually it would be 1999)

By ensuring there is a space character right after the chapter number
(as in "1 title") we make sure to avoid this get the right chapter.