Added novel specific cleanup function
I've noticed that some novels need some kind of novel specific cleaning,
and figured it was a good idea to allow the same for all of them.

The first way to implement this that came to mind was to unconditionally
call a "module.clean()" function on every single chapter.

Those novels that need this will implement a way to clean themselves,
those which don't will make just do nothing.